add :added the envfile to pubspec and linked packages for access add : loaded envfile in main add : authentification in sensordata and subscribes to changes from the pocketbase server DOESNT WORK YET (to fix) edit : minor bug correction
/!\ EVERYONE NEEDS TO change the envfile password line by adding ' around the password ('password')

+++ tristapp/lib/page/mappage.dart
@@ -47,8 +47,8 @@ class _MapPageState extends State<MapPage> {
               height: 30,
               point: LatLng(lat, long),
               child: MapPinInfo(
-                lat: (lat.toInt()),
-                long: (long.toInt()),
+                lat: (lat),
+                long: (long),
                 //key: stationName,
               ),
             ),
